Move silx.gui.utils._image to silx.gui.utils.image
