+++ silx/gui/plot/actions/io.py
@@ -53,7 +53,7 @@ from silx.gui import qt, printer
 from silx.gui.dialog.GroupDialog import GroupDialog
 from silx.third_party.EdfFile import EdfFile
 from silx.third_party.TiffIO import TiffIO
-from ...utils._image import convertArrayToQImage
+from ...utils.image import convertArrayToQImage
 if sys.version_info[0] == 3:
     from io import BytesIO
 else:
